web/requirements.txt
Flask==2.3.3
gunicorn==20.1.0
redis==4.6.0
python-dotenv==1.0.0
docker==6.1.3